GCS calcRelativeLidarPose  done! (matlab compared)

+++ include/util/sophus_lie.hpp
@@ -11,7 +11,7 @@
 #include "Eigen/Geometry"
 
 namespace sophuslie {
-	void se3Exp(const Eigen::Matrix<float,6,1>& xi, Eigen::Matrix4d& T);
+	void se3Exp(const Eigen::Matrix<double,6,1>& xi, Eigen::Matrix4d& T);
 	void se3Exp(const Eigen::MatrixXd& xi_, Eigen::MatrixXd& T_);
 	void se3Exp(const Eigen::Matrix<float,6,1>& xi_, Eigen::Matrix4f& T_);
 	void SE3Log(const Eigen::MatrixXf& T, Eigen::MatrixXf& xi);
